+++ CHANGELOG.md
@@ -5,7 +5,7 @@ dd mmm yyyy
 
 - New: error limit for apps.
 - Layout fixes.
-- Added player ID to account name.
+- Added player ID to account name everywhere.
 
 ## 0.7.0